+++ core/src/main/java/xyz/v2my/easymodeling/randomizer/numberrandomizer/DoubleRandomizer.java
@@ -6,7 +6,7 @@ import xyz.v2my.easymodeling.randomizer.NumberRandomizer;
 public class DoubleRandomizer extends NumberRandomizer<Double> {
 
     @Override
-    public Double next(long min, long max) {
+    public Double next(double min, double max) {
         return doubleBetween(min, max);
     }
 }
