+++ core/src/main/java/xyz/v2my/easymodeling/randomizer/numberrandomizer/IntegerRandomizer.java
@@ -6,7 +6,7 @@ import xyz.v2my.easymodeling.randomizer.NumberRandomizer;
 public class IntegerRandomizer extends NumberRandomizer<Integer> {
 
     @Override
-    public Integer next(long min, long max) {
+    public Integer next(double min, double max) {
         return doubleBetween(min, max).intValue();
     }
 }
